backend/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv

# Data Processing
pandas==2.2.3
pyarrow==18.1.0
openpyxl==3.1.5
xlrd==2.0.1
duckdb

# Detection & Parsing
charset-normalizer==3.4.1
python-magic-bin==0.4.14
clevercsv==0.8.2
xmltodict==0.14.1
sqlparse==0.5.3
ijson==3.3.0

# System Monitoring
psutil==6.1.1

# Statistical Analysis & ML
scipy==1.14.1
scikit-learn==1.6.1
statsmodels==0.14.4

# NLP & Text
nltk
textblob
langdetect
jellyfish

# LLM
google-generativeai

# Reporting
python-docx
fpdf2

# Rate Limiting
slowapi

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
